requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
groq==0.4.1
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
PyMuPDF==1.23.8
pypdf==3.17.0
llama-index==0.9.15